[Feat] content 모듈화
