+++ feature/content/detail/src/commonMain/kotlin/com/whatever/caramel/feature/content/detail/empty.kt
@@ -1,0 +1,10 @@
+package com.whatever.caramel.feature.content.detail
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun ContentDetail(
+
+){
+
+}
