+++ feature/content/edit/src/commonMain/kotlin/com/whatever/caramel/feature/content/edit/empty.kt
@@ -1,0 +1,10 @@
+package com.whatever.caramel.feature.content.edit
+
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun ContentEdit(
+
+){
+
+}
