--- feature/content/src/commonMain/kotlin/com/whatever/caramel/feature/content/di/Module.kt
@@ -1,9 +1,0 @@
-package com.whatever.caramel.feature.content.di
-
-import com.whatever.caramel.feature.content.ContentViewModel
-import org.koin.core.module.dsl.viewModelOf
-import org.koin.dsl.module
-
-val contentFeatureModule = module {
-    viewModelOf(::ContentViewModel)
-}
