--- feature/content/src/commonMain/kotlin/com/whatever/caramel/feature/content/mvi/ContentSideEffect.kt
@@ -1,10 +1,0 @@
-package com.whatever.caramel.feature.content.mvi
-
-import com.whatever.caramel.core.viewmodel.UiSideEffect
-
-sealed interface ContentSideEffect : UiSideEffect {
-
-    data object NavigateToBackStack : ContentSideEffect
-
-    data class ShowErrorSnackBar(val code: String, val message: String? = null) : ContentSideEffect
-}
