+++ settings.gradle.kts
@@ -44,7 +44,9 @@ include(":core:deeplink")
 include(":feature")
 include(":feature:splash")
 include(":feature:login")
-include(":feature:content")
+include(":feature:content:create")
+include(":feature:content:edit")
+include(":feature:content:detail")
 include(":feature:couple")
 include(":feature:main")
 include(":feature:setting")
@@ -56,7 +58,6 @@ include(":feature:home")
 include(":feature:memo")
 include(":feature:couple:connect")
 include(":feature:couple:connecting")
-include(":feature:content")
 include(":feature:couple:invite")
 include(":core:testing")
 include(":core:util")
